refactor(deployments): extract render apply pipeline
Refs: HOL-1099
