+++ internal/deploymentrender/render_raw_cue.go
@@ -1,0 +1,102 @@
+package deploymentrender
+
+import (
+	"context"
+	"fmt"
+
+	"cuelang.org/go/cue"
+	"cuelang.org/go/cue/cuecontext"
+
+	v1alpha2 "github.com/holos-run/holos-console/api/v1alpha2"
+)
+
+// EvaluateGroupedCUE compiles and evaluates a pre-concatenated CUE source
+// document (template + any already-embedded raw CUE input) and returns the
+// rendered Kubernetes resources grouped by origin. This is the raw-CUE entry
+// point used by callers that assemble the full CUE document themselves —
+// for example, the templates preview path which receives CUE strings for
+// "platform" and "input" from the client.
+//
+// When readPlatformResources is true the renderer reads both
+// platformResources and projectResources; when false only projectResources is
+// read (per ADR 016 Decision 8 the project-level path must not emit
+// platformResources).
+//
+// This helper is a transitional raw-CUE entry point kept out of render.go so
+// that render.go exposes exactly one public render function
+// (CueRenderer.Render). The raw-CUE call sites in the templates package will
+// be removed in later phases (see HOL-562); when the last caller is gone,
+// this file and the helper it exports will be deleted.
+func EvaluateGroupedCUE(ctx context.Context, combinedCUESource string, readPlatformResources bool) (*GroupedResources, error) {
+	evalCtx, cancel := context.WithTimeout(ctx, renderTimeout)
+	defer cancel()
+
+	type result struct {
+		grouped *GroupedResources
+		err     error
+	}
+	ch := make(chan result, 1)
+	go func() {
+		grouped, err := evaluateCueInput(combinedCUESource, readPlatformResources)
+		ch <- result{grouped, err}
+	}()
+
+	select {
+	case <-evalCtx.Done():
+		return nil, fmt.Errorf("CUE template evaluation timed out after %s", renderTimeout)
+	case res := <-ch:
+		return res.grouped, res.err
+	}
+}
+
+// evaluateCueInput performs synchronous CUE template evaluation of a
+// pre-concatenated source document. The caller is responsible for assembling
+// the full CUE document (template plus any raw-CUE "platform" / "input"
+// values). Generated schema definitions are prepended so templates can
+// reference #PlatformInput, #ProjectInput, etc.
+//
+// At least one of projectResources.namespacedResources or
+// platformResources.namespacedResources must exist (preview of a
+// platform-only template is permitted).
+func evaluateCueInput(cueSource string, readPlatformResources bool) (*GroupedResources, error) {
+	cueCtx := cuecontext.New()
+
+	combined := v1alpha2.GeneratedSchema + "\n" + cueSource
+	unified := cueCtx.CompileString(combined)
+	if err := unified.Err(); err != nil {
+		return nil, fmt.Errorf("invalid CUE template: %w", err)
+	}
+
+	namespacedValue := unified.LookupPath(cue.ParsePath("projectResources.namespacedResources"))
+	platformNamespacedValue := unified.LookupPath(cue.ParsePath("platformResources.namespacedResources"))
+
+	// Three-way distinction:
+	//   1. Neither path exists → template is missing the required
+	//      structured-output fields; keep the existing message.
+	//   2. At least one path exists but carries a CUE evaluation error (e.g.
+	//      non-concrete dynamic field key) → surface the CUE error verbatim so
+	//      template authors can act on the real cause.
+	//   3. At least one path exists with no error → happy path.
+	//
+	// Note: LookupPath on a missing field returns a value that is both
+	// !Exists() AND has a non-nil Err(), so we must test Exists() first before
+	// inspecting Err() — Err() alone cannot distinguish "absent" from "broken".
+	projExists := namespacedValue.Exists()
+	platExists := platformNamespacedValue.Exists()
+
+	if !projExists && !platExists {
+		return nil, fmt.Errorf("template must define 'projectResources.namespacedResources' or 'platformResources.namespacedResources' (structured output format required)")
+	}
+	if projExists {
+		if err := namespacedValue.Err(); err != nil {
+			return nil, fmt.Errorf("projectResources.namespacedResources: %w", err)
+		}
+	}
+	if platExists {
+		if err := platformNamespacedValue.Err(); err != nil {
+			return nil, fmt.Errorf("platformResources.namespacedResources: %w", err)
+		}
+	}
+
+	return evaluateStructuredGrouped(unified, readPlatformResources)
+}
